# api-management/api-documentation.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Documentation API du projet money
  description: Décrit la structure de l'API du projet money de l'équipe NALT
servers:
    - url: http://localhost:3000/api
      description: Serveur de développement de l'API
paths:
...
  /utilisateurs:
    # request get for the path /utilisateurs
    get:
      summary: Récupère la liste de tous les utilisateurs
      security:
        - bearerAuth: []
      tags:
        - Utilisateurs
      description: Retourne la liste des utilisateurs avec l'id, le nom, le prenom, le login, la ville, le code postal et la date de création
      responses:
        '200':
          $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    # request post for the path /utilisateurs
    post:
      summary: Créé un nouvel utilisateur.
      security:
        - bearerAuth: []
      tags:
        - Utilisateurs
      description: Créé un nouvel utilisateur avec le nom, le prenom, le mot de passe, la ville et le code postal.
      requestBody: # specification of the request body
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nomUtilisateur:
                  type: string
                prenomUtilisateur:
                  type: string
                motDePasse:
                  type: string
                  format: password
                ville:
                  type: string
                codePostal: 
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /utilisateurs/{idUtilisateur}:
    # request get fot the path /utilisateurs/{idUtilisateur}
    get:
      summary: Récupère un utilisateur par son id.
      security:
        - bearerAuth: []
      tags:
        - Utilisateurs
      description: Récupère un utilisateur avec l'id, le nom, le prenom, le login, la ville et le code postal.
      parameters: # parameters for the query
        - name: idUtilisateur
          description: Id de l'utilisateur
          in: path
          required: true
          schema:
            type: integer
            
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    # request put for the path /utilisateurs/{idUtilisateur}
    put:
      summary: Met à jour les informations d'un utilisateur.
      security:
        - bearerAuth: []
      tags:
        - Utilisateurs
      description: Met à jour plusieurs informations d'un utilisateur comme le nom, le prenom, le mot de passe, le login, la ville et le code postal.

      requestBody: # specification of the request body
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                nomUtilisateur:
                  type: string
                prenomUtilisateur:
                  type: string
                motdepasse:
                  type: string
                  format: password
                ville:
                  type: string
                codePostal: 
                  type: string

      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  
    #request patch for the path /utilisateurs/{idUtilisateur}
    patch:
      summary: Met à jour partiellement les informations d'un utilisateur.
      security:
        - bearerAuth: []
      tags:
        - Utilisateurs
      description: Met à jour partiellement les informations d'un utilisateur comme le nom, le prenom, le mot de passe, le login, la ville et le code postal.
      parameters: # parameters for the query
        - name: idUtilisateur
          description: Id de l'utilisateur
          in: path
          required: true
          schema:
            type: integer

      requestBody: # specification of the request body
        required: false
        content:
          application/json:
            schema:
              type: object
            properties:
                nomUtilisateur:
                  type: string
                prenomUtilisateur:
                  type: string
                motdepasse:
                  type: string
                  format: password
                ville:
                  type: string
                codePostal: 
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    # request delete for the path /utilisateurs/{idUtilisateur}
    delete:
      summary: Supprime un utilisateur par son id
      security:
        - bearerAuth: []
      tags:
        - Utilisateurs
      description: Met à jour partiellement les informations d'un utilisateur comme le nom, le prenom, le mot de passe, le login, la ville et le code postal.
      parameters: # parameters for the query
        - name: idUtilisateur
          description: Id de l'utilisateur
          in: path
          required: true
          schema:
            type: integer

      responses:
        '200':
          $ref: '#/components/responses/Success'
        '204':
          $ref: '#/components/responses/NoContent'
        '303':
          $ref: '#/components/responses/SeeOther'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
  # subroute /comptes
...
components:
...
  schemas:    
...
    Error:
      description: Error response
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    
    # schema for the success responses
    Success:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          additionalProperties: true
      
  # responses
  responses:
...
    Success:
      description: Success
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Success'
    # 201 Created
    Created:
      description: Created
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Success'
    # 204 No Content
    NoContent:
      description: No Content
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Success'
    # 303 See Other
    SeeOther:
      description: See Other
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Success'     
    # 400 Bad Request
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 401 Unauthorized
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 403 Forbidden
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 404 Not Found
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'  
    # 406 Not Acceptable
    NotAcceptableError:
      description: Not Acceptable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 409 Conflict
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 500 Internal Server Error
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # securitySchemes
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# security
security:
  - bearerAuth: []